src/backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
build==1.2.2.post1
CacheControl==0.14.0
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.3.0
cleo==2.1.0
click==8.3.0
crashtest==0.4.1
cryptography==41.0.4
distlib==0.3.8
distro==1.9.0
dulwich==0.24.2
fastapi==0.119.1
fastjsonschema==2.18.1
filelock==3.16.1
findpython==0.6.2
greenlet==3.2.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.4
installer==0.7.0
jaraco.classes==3.3.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.6.0
langchain==1.0.5
langchain-core==1.0.4
langchain-openai==1.0.2
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.42
more-itertools==10.1.0
msgpack==1.1.2
openai==2.7.2
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pbs-installer==2025.2.12
pipenv==2025.0.4
pkginfo==1.12.0
platformdirs==4.3.6
poetry==2.2.1
poetry-core==2.2.1
pycparser==2.21
pydantic==2.12.3
pydantic_core==2.41.4
pyproject_hooks==1.0.0
python-dotenv==1.2.1
PyYAML==6.0.3
RapidFuzz==3.14.1
regex==2025.11.3
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
setuptools==80.9.0
shellingham==1.5.3
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.48.0
tenacity==9.1.2
tiktoken==0.12.0
tomlkit==0.12.1
tqdm==4.67.1
trove-classifiers==2023.9.19
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.4
xxhash==3.6.0
zstandard==0.23.0
